% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag.R
\name{bag_geocode}
\alias{bag_geocode}
\title{Retrieve the (RD) coordinates of a street address in The Hague}
\usage{
bag_geocode(postcode, huisnummer, straatnaam, plaatsnaam, max.retries = 2L)
}
\arguments{
\item{postcode}{a character verctor with postal code}

\item{huisnummer}{a character vector with house numbers}

\item{straatnaam}{a character vector with street names}

\item{plaatsnaam}{a character vector with city or location name}

\item{max.retries}{an integer with the maximum number of retries if a request
times out (default: 2)}
}
\value{
A data frame with the coordinates of each address in two columns
  named \sQuote{rdx} and \sQuote{rdy}. Unknown addresses have NA as
  coordinates.
}
\description{
Retrieve the (RD) coordinates of a street address in The Hague
}

