% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir.R
\name{cir_getCase}
\alias{cir_getCase}
\title{Retrieve case details}
\usage{
cir_getCase(publication.nr)
}
\arguments{
\item{publication.nr}{a character vector with publication numbers}
}
\value{
A named list with the details for each publication in
  \code{publicationNumber}.
}
\description{
Retrieve case details
}

