% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir.R
\name{cir_getCourtNames}
\alias{cir_getCourtNames}
\title{Get a vector of court names/locations}
\usage{
cir_getCourtNames()
}
\value{
A character vector with the available courts in the Netherlands.
}
\description{
Get a vector of court names/locations
}

