% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir.R
\name{cir_GetLastUpdate}
\alias{cir_GetLastUpdate}
\title{Retrieve the date of the last update}
\usage{
cir_GetLastUpdate()
}
\value{
A \code{Date} object with the date of the last update to the Dutch
  insolvency registry.
}
\description{
Retrieve the date of the last update
}

