% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir.R
\name{cir_searchByDate}
\alias{cir_searchByDate}
\title{Retrieve publication numbers by date}
\usage{
cir_searchByDate(date, court, type = NULL)
}
\arguments{
\item{date}{a \code{Date} object which should not be in the future, nor more 
than one year in the past}

\item{court}{an optional character string with the name of the court, e.g. 
\sQuote{Den Haag}}

\item{type}{a character vector with publication types to retrieve, accepted
values are \sQuote{Uitspraken faillissement}, \sQuote{Uitspraken
schuldsanering}, \sQuote{Uitspraken surseance}, \sQuote{Einde
faillissementen}, \sQuote{Einde schuldsaneringen}, \sQuote{Einde
surseances} and \sQuote{Overig}. If no value is supplied then all
publication types are retrieved.}
}
\value{
A character vector with case names or \code{NULL} if no cases are 
  found or some other error has occurred.
}
\description{
Retrieve publication numbers by date
}
\details{
The function name \sQuote{searchByDate} is actually quite a misnomer
  as the function in fact returns all publications since the date provided.
}
\seealso{
\code{\link{cir_getCourtNames}} to get a vector of court names
}

