% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epbd.R
\name{epbd_downloadCertificates}
\alias{epbd_downloadCertificates}
\title{Download energy certificates at the start of the current month}
\usage{
epbd_downloadCertificates(postcode.min = 1000L, postcode.max = 9999L)
}
\arguments{
\item{postcode.min}{integer value representing the smallest postcode to 
download (default: 1000)}

\item{postcode.max}{integer value representing the largest postcode to
download (default: 9999); must be greater than \code{postcode.min}}
}
\value{
A data frame with the contents of the energy certificates.
}
\description{
Download energy certificates at the start of the current month
}
\details{
The registry with \sQuote{energieprestatielabels} uses a webservice 
  based on the SOAP protocol. There are two types of queries which you can 
  send to the webservice: \itemize{ \item retrieve all existing certificates 
  at the beginning of the month or \item retrieve the mutations (i.e. 
  changes, updates, etc.) on a given date. } This function implements the 
  first query: retrieve all certificates at the start of the month. The 
  arguments \code{postcode.min} and \code{postcode.max} can be used to 
  download certificates from a specific region. These arguments accept only 
  the 4 digit numeric part of Dutch postal codes.
  
  Unfortunately, the webservice doesn't actually return an XML document with 
  the mutations, it returns an XML document with only a link to the (zipped) 
  XML document on the server. So this function first queries the webservice 
  to retrieve the URL to the archive, then downloads and extracts the 
  archive, and finally reads the contents into an \code{xml_document} object.
}
\note{
This functions calls a Python script in the \sQuote{inst/tools} folder 
  which uses a SAX-based XML parser to extract the certificate data from the 
  huge XML file which is returned by this rather stupid service.
}
\references{
The registry can be accessed online at 
  \url{http://www.ep-online.nl/}.
}

