% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epbd.R
\name{epbd_extractLabelMutations}
\alias{epbd_extractLabelMutations}
\title{Extract details of EP label mutations}
\usage{
epbd_extractLabelMutations(xml)
}
\arguments{
\item{xml}{a character string with the name of an XML file containing changes
(i.e. mutations) in the EP-label registry or an object of class
\code{xml_document} with the same data already read from the file system}
}
\value{
A data frame with a selection of fields extracted from the mutations
  file.
}
\description{
Extract details of EP label mutations
}

