% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epbd.R
\name{epbd_getLabelMutations}
\alias{epbd_getLabelMutations}
\title{Retrieve label mutations}
\usage{
epbd_getLabelMutations(date)
}
\arguments{
\item{date}{a Date object}
}
\value{
An object of class \code{xml_document} which you can pass to 
  \code{\link{epbd_extractLabelMutations}} to convert to a data frame.
}
\description{
Retrieve label mutations
}
\details{
The registry with \sQuote{energieprestatielabels} uses a webservice 
  based on the SOAP protocol. There are two types of queries which you can 
  send to the webservice: \itemize{ \item retrieve all existing labels at the
  beginning of the month or \item retrieve the mutations (i.e. changes, 
  updates, etc.) on a given date. } This function implements the second 
  query: retrieve all changes on a given date.
  
  Unfortunately, the webservice doesn't actually return an XML document with
  the mutations, it returns an XML document with only a link to the (zipped)
  XML document on the server. So this function first queries the webservice
  to retrieve the URL to the archive, then downloads and extracts the
  archive, and finally reads the contents into an \code{xml_document} object.
}
\note{
The webservice only provides access to the information for the current 
  month. In other words: it is not possible to retrieve mutations in other 
  months except for the current.
}
\references{
The registry can be accessed online at \url{http://www.ep-online.nl/}.
}

