% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epbd.R
\name{epbd_labelColor}
\alias{epbd_labelColor}
\title{Energy label HEX color code}
\usage{
epbd_labelColor(label)
}
\arguments{
\item{a}{character vector with energy label codes (i.e. \sQuote{A}, 
\sQuote{B}, etc.)}
}
\value{
A character vector with HEX color codes. \code{NA} values are given a
  light grey color.
}
\description{
Energy label HEX color code
}

