% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epbd.R
\name{epbd_login}
\alias{epbd_login}
\title{Provide login credentials for the energy label registry}
\usage{
epbd_login(username, password)
}
\arguments{
\item{username}{a single character string with the username}

\item{password}{a single character string with the password}
}
\value{
\code{TRUE} (invisibly) in case of success
}
\description{
Provide login credentials for the energy label registry
}
\details{
Register your login credentials once at the start of your session
and all subsequent queries to the registry will use these.
}

